refs: #23 FAQコンポーネントを作成
